split wrapper from editor move config to runner extract apple_watch identification in runner
