+++ frameit/lib/frameit.rb
@@ -9,5 +9,6 @@ require 'frameit/template_finder'
 require 'frameit/mac_editor'
 require 'frameit/dependency_checker'
 require 'frameit/options'
+require 'frameit/wrapper/wrapper'
 require 'frameit/wrapper/trim_box'
 require 'frameit/wrapper/strings_parser'
